refactor: reorganize types to be reused
